WB criterion, OrElse, Prover9
Added OrElse language optimisation and modified the OptSen translate mode so that it also takes well-behavedness of the output formula into account. Also added CheckAgain.py, which should replace EquivalenceChecker.py. This new equivalence checker makes use of the external prover Prover9, and is able to prove a whole bunch of equivalences. Also added linearisation for PTaut and PContra in both English and GGC. This helps the prover as well.